setting up tests with pytest fixtures
