+++ tests/conftest.py
@@ -16,6 +16,6 @@ sys.path.insert(0, str(TESTS_DIR_PARENT))
 
 # module import paths to python files containing fixtures
 pytest_plugins = [
-    # e.g. "tests/fixtures/example_fixture.py" should be registered as:
-    "tests.fixtures.example_fixture",
+    # "tests/fixtures/mocked_aws.py" should be registered as:
+    "tests.fixtures.mocked_aws",
 ]
