+++ tests/fixtures/mocked_aws.py
@@ -1,0 +1,43 @@
+"""Pytest fixture to mock AWS services."""
+
+import os
+from typing import Generator, Any
+
+import boto3
+import pytest
+from moto import mock_aws
+
+from tests.consts import TEST_BUCKET_NAME
+
+
+# Set the environment variables to point away from AWS
+def point_away_from_aws() -> None:
+    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
+    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
+    os.environ["AWS_SECURITY_TOKEN"] = "testing"
+    os.environ["AWS_SESSION_TOKEN"] = "testing"
+    os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
+
+
+# Our fixture is a function and we have named it as a noun instead
+# of verb, because it is a resource that is being provided to the test.
+
+
+@pytest.fixture(scope="function")
+def mocked_aws() -> Generator[None, Any, None]:
+    with mock_aws():
+        # Set the environment variables to point away from AWS
+        point_away_from_aws()
+
+        # 1. Create an S3 bucket
+        s3_client = boto3.client("s3")
+        s3_client.create_bucket(Bucket=TEST_BUCKET_NAME)
+
+        yield
+
+        # 4. Clean up/Teardown by deleting the bucket
+        response = s3_client.list_objects_v2(Bucket=TEST_BUCKET_NAME)
+        for obj in response.get("Contents", []):
+            s3_client.delete_object(Bucket=TEST_BUCKET_NAME, Key=obj["Key"])
+
+        s3_client.delete_bucket(Bucket=TEST_BUCKET_NAME)
